+++ NEWS.md
@@ -1,6 +1,7 @@
 # swissknife 0.36
 
-* Add .assertPackagesAvailable utility function and unit test
+* Add .assertPackagesAvailable utility function and unit test to templates
+* Update internally used .assertPackagesAvailable to be identical to the ones in templates
 
 # swissknife 0.35